Update patch set 1
Patch Set 1:

Well, C6606 is the U.S. variant, so that's ruled out.
C6603 is the European LTE variant, which is widely available.
I don't have the fingerprint of the latest build of C6602, so I just kept it the same (removed/5), and updated it to .434 in another commit.

Patch-set: 1