Update patch set 1
Patch Set 1: Code-Review-2

World writable is just bad, and over that you're doing that to a bunch of unnecessary files.

It got broken due to a fixing a security risk that Torch was piggy-backing on, but this is not the solution.

Patch-set: 1
Reviewer: Chirayu Desai <2195@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=-2