Update patch set 2
Patch Set 2:

Mass storage mode isn't available in the standard Android UI, which is what CTS covers. It does not cover cases where a user gains root and switches the usb mode on the command line, or uses an app that most likely requires root to expose UMS. 

Therefore, even if he did enable UMS in the xml, this patch could not break CTS, since UMS is still not available in the standard UI to a non-root user.

Patch-set: 2